requirements.txt
streamlit>=1.35
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
plotly>=5.20
python-dateutil>=2.9
scipy>=1.11
joblib>=1.3
tensorflow==2.19.0   # required for Autoencoder (if TF is not needed, remove this line)
